cleaned up various tests, fixed rules, refactored tests so all run on test2
