+++ j2/rules/Cliente_SAP_PosVenda_POLARIS.yaml
@@ -3,37 +3,37 @@ updatereq:
   vertices:
     
     - label: Person.Natural
-      condition: ${Tax_Number_1?.trim()?.length() == 11 || Tax_Number_2?.trim()?.length() == 11}
+      condition: ${Tax_Number_1?.replaceAll('[^0-9]', '')?.length() == 11 || Tax_Number_2?.replaceAll('[^0-9]', '')?.length() == 11}
       props:
         - name: Person.Natural.Full_Name
           val: ${Name_1?.trim()?.toUpperCase()}${Name_2?.trim()?' ':''}${Name_2?.trim()?.toUpperCase()}
         - name: Person.Natural.Customer_ID
-          val: ${Tax_Number_1?.trim()?.length() == 11?Tax_Number_1:Tax_Number_2}
+          val: ${Tax_Number_1?.replaceAll('[^0-9]', '')?.length() == 11?Tax_Number_1?.replaceAll('[^0-9]', ''):Tax_Number_2?.replaceAll('[^0-9]', '')}
         - name: Person.Natural.Type
           val: Clientes
           excludeFromSearch: True
     
     - label: Object.Identity_Card
-      condition: ${Tax_Number_1?.trim()?.length() == 11 || Tax_Number_2?.trim()?.length() == 11}
+      condition: ${Tax_Number_1?.replaceAll('[^0-9]', '')?.length() == 11 || Tax_Number_2?.replaceAll('[^0-9]', '')?.length() == 11}
       props:
         - name: Object.Identity_Card.Id_Value
-          val: ${Tax_Number_1?.trim()?.length() == 11?Tax_Number_1:Tax_Number_2}
+          val: ${Tax_Number_1?.replaceAll('[^0-9]', '')?.length() == 11?Tax_Number_1?.replaceAll('[^0-9]', ''):Tax_Number_2?.replaceAll('[^0-9]', '')}
           mandatoryInSearch: True
         - name: Object.Identity_Card.Id_Type
           val: CPF
           excludeFromSearch: True
     
     - label: Person.Organisation
-      condition: ${Tax_Number_1?.trim()?.length() == 14 || Tax_Number_2?.trim()?.length() == 14}
+      condition: ${Tax_Number_1?.replaceAll('[^0-9]', '')?.length() == 14 || Tax_Number_2?.replaceAll('[^0-9]', '')?.length() == 14}
       props:
         - name: Person.Organisation.Registration_Number
-          val: ${Tax_Number_1?.trim()?.length() == 14?Tax_Number_1:Tax_Number_2}
+          val: ${Tax_Number_1?.replaceAll('[^0-9]', '')?.length() == 14?Tax_Number_1?.replaceAll('[^0-9]', ''):Tax_Number_2?.replaceAll('[^0-9]', '')}
           mandatoryInSearch: True
         - name: Person.Organisation.Name
           val: ${Name_1?.trim()?.toUpperCase()} ${Name_2?.trim()?.toUpperCase()}
     
     - label: Person.Natural
-      condition: ${!(Tax_Number_1?.trim()?.length() == 11 || Tax_Number_2?.trim()?.length() == 11 || Tax_Number_1?.trim()?.length() == 14 || Tax_Number_2?.trim()?.length() == 14) || (Tax_Number_1?.trim()?.matches('.*[a-zA-Z]+.*') || Tax_Number_2?.trim()?.matches('.*[a-zA-Z]+.*'))}
+      condition: ${!(Tax_Number_1?.replaceAll('[^0-9]', '')?.length() == 11 || Tax_Number_2?.replaceAll('[^0-9]', '')?.length() == 11 || Tax_Number_1?.replaceAll('[^0-9]', '')?.length() == 14 || Tax_Number_2?.replaceAll('[^0-9]', '')?.length() == 14)}
       props:
         - name: Person.Natural.Full_Name
           val: ${Name_1?.trim()?.toUpperCase()}${Name_2?.trim()?' ':''}${Name_2?.trim()?.toUpperCase()}
