+++ j2/rules/sharepoint_treinamentos.yaml
@@ -46,27 +46,33 @@ updatereq:
       props:
         - name: Event.Group_Ingestion.Metadata_Start_Date
           val: ${new Date()}
-          mandatoryInSearch: True
           type: "java.util.Date"
+          excludeFromSearch: True
         - name: Event.Group_Ingestion.Metadata_End_Date
           val: ${new Date()}
           type: "java.util.Date"
           excludeFromSearch: True
         - name: Event.Group_Ingestion.Type
           val: sharepoint/treinamentos
-          excludeFromSearch: True
+          mandatoryInSearch: True
         - name: Event.Group_Ingestion.Operation
           val: Structured Data Insertion
-          excludeFromSearch: True
+          mandatoryInSearch: True
+        - name: Event.Group_Ingestion.Ingestion_Date
+          val: ${new java.text.SimpleDateFormat('yyyy-MM-dd').format(new Date())}
+          mandatoryInSearch: True
     
     - label: Object.Data_Source
       props:
         - name: Object.Data_Source.Name
           val: sharepoint/treinamentos
+          mandatoryInSearch: True
         - name: Object.Data_Source.Type
           val: sharepoint/treinamentos
+          mandatoryInSearch: True
         - name: Object.Data_Source.Description
           val: Registros de Treinamentos LGPD Colaboradores
+          mandatoryInSearch: True
 
   edges:
   - label: Course_Training
