+++ src/test/java/com/pontusvision/gdpr/AppTest2.java
@@ -1,15 +1,28 @@
 package com.pontusvision.gdpr;
 
 import com.google.gson.Gson;
+import com.google.gson.JsonObject;
+import com.google.gson.JsonParser;
+import com.orientechnologies.apache.commons.csv.CSVFormat;
+import com.orientechnologies.apache.commons.csv.CSVRecord;
 import com.pontusvision.ingestion.Ingestion;
 import com.pontusvision.ingestion.IngestionCSVFileRequest;
 import org.apache.commons.io.FileUtils;
+import org.apache.commons.io.IOUtils;
+import org.apache.http.client.HttpClient;
+import org.apache.http.client.methods.HttpGet;
+import org.apache.http.client.methods.HttpPost;
+import org.apache.http.entity.StringEntity;
+import org.apache.http.impl.client.HttpClients;
 import org.junit.AfterClass;
 import org.junit.BeforeClass;
 import org.junit.Test;
 
 import java.io.File;
+import java.io.FileReader;
 import java.io.IOException;
+import java.io.Reader;
+import java.net.URI;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.*;
@@ -33,50 +46,8 @@ public class AppTest2  extends AppTest{
    */
 
 
-
-  @AfterClass
-  public static void after() throws Exception {
-//    App.gserver.stop().join();
-    App.oServer.shutdown();
-    server.stop();
-
-  }
-  public static boolean deleteDirectory(File directoryToBeDeleted) {
-    File[] allContents = directoryToBeDeleted.listFiles();
-    if (allContents != null) {
-      for (File file : allContents) {
-        deleteDirectory(file);
-      }
-    }
-    return directoryToBeDeleted.delete();
-  }
-
-  @BeforeClass
-  public static void before() throws Exception {
-    Path resourceDirectory = Paths.get(".");
-    String absolutePath = resourceDirectory.toFile().getAbsolutePath();
-
-    String databaseDir = Paths.get(absolutePath, "databases").toString();
-    deleteDirectory(new File(databaseDir));
-
-    String jpostalDataDir = Paths.get(absolutePath, "jpostal", "libpostal").toString();
-    System.setProperty("user.dir", absolutePath);
-    System.setProperty("ORIENTDB_ROOT_PASSWORD", "pa55word");
-    System.setProperty("ORIENTDB_HOME", absolutePath);
-    System.setProperty("pg.jpostal.datadir", jpostalDataDir);
-
-    server = App.createJettyServer();
-
-    server.start();
-    App.init(Paths.get(absolutePath, "config", "gremlin-server.yaml").toString());
-
-     System.out.println("finished Init");
-  }
-
-  static Gson gson = new Gson();
-
   @Test
-  public void test3() throws InterruptedException {
+  public void testPhase1CSVIngestion() throws InterruptedException {
     try {
 
         csvTestUtil("phase1.csv",  "phase1_csv");
@@ -106,7 +77,7 @@ public class AppTest2  extends AppTest{
   public void testCsvPolaris() throws InterruptedException {
     try {
 
-      csvTestUtil("clientes.csv",  "Cliente_SAP_PosVenda_POLARIS");
+      csvTestUtil("sap-polaris-clientes.csv",  "Cliente_SAP_PosVenda_POLARIS");
 
 //    Testing for Person.Natural WITH Tax_Number
       String userId0 =
@@ -115,19 +86,20 @@ public class AppTest2  extends AppTest{
       String pintoConnectionsQuery = "App.g.V(\"" + userId0 +"\").bothE().count().next().toString()";
       String pintoConnections = App.executor.eval(pintoConnectionsQuery)
               .get().toString();
-      assertEquals(pintoConnections,"4");
+      assertEquals("5", pintoConnections); // 2 phone numbers + 1 identity card + 1 location + 1 ingestion event
 
 //    Testing for Person.Natural withOUT Tax_Number
-      String userId1 =
+      String munirConnections =
               App.executor.eval("App.g.V().has('Person.Natural.Full_Name', eq('MUNIR HANDAUS'))\n" +
-                      ".next().id().toString()").get().toString();
+                      ".bothE().count().next().toString()").get().toString();
 //      String munirConnectionsQuery = "App.g.V(\"" + userId1 +"\").bothE().count().next().toString()";
 //      String munirConnections = App.executor.eval(munirConnectionsQuery).get().toString();
+      assertEquals("3", munirConnections); // 1 phone number + 0 identity card (sem tax) + 1 location + 1 ingestion event
 
      String personNaturalTypes =
              App.executor.eval("App.g.V().has('Person.Natural.Type', eq('Clientes Sem Tax'))" +
                      ".count().next().toString()").get().toString();
-     assertEquals(personNaturalTypes, "5");
+     assertEquals("5",personNaturalTypes);
 
      System.out.println("*******************************************************\n" + personNaturalTypes);
 
@@ -167,13 +139,13 @@ public class AppTest2  extends AppTest{
                       ".next().id().toString()").get().toString();
 
 //    test for Object.Data_Source.Name
-      String userId3 =
+      String sharepointDataSourceId =
               App.executor.eval("App.g.V().has('Object.Data_Source.Name', eq('sharepoint/treinamentos'))" +
                       ".next().id().toString()").get().toString();
 //      assertEquals(userId3,"#82:0"); -- Id can change!!
-      String rootConnectionsQuery = "App.g.V(\"" + userId3 +"\").bothE().count().next().toString()";
+      String rootConnectionsQuery = "App.g.V(\"" + sharepointDataSourceId +"\").bothE().count().next().toString()";
       String rootConnections = App.executor.eval(rootConnectionsQuery).get().toString();
-      assertEquals(rootConnections,"3");
+      assertEquals("1",rootConnections);
 
 //    test for Object.Awareness_Campaign.Description
       String userId4 =
@@ -181,7 +153,7 @@ public class AppTest2  extends AppTest{
                  "eq('1º CURSO - LGPD - LEI GERAL DE PROTEÇÃO DE DADOS')).next().id().toString()").get().toString();
       String curso1ConnectionsQuery = "App.g.V(\"" + userId4 +"\").bothE().count().next().toString()";
       String curso1Connections = App.executor.eval(curso1ConnectionsQuery).get().toString();
-      assertEquals(curso1Connections,"2"); // -- number of Group.Ingestion nodes can vary !!
+      assertEquals(curso1Connections,"2");
 
 //    test for Object.Awareness_Campaign.Description 2
       String userId5 =
@@ -217,7 +189,7 @@ public class AppTest2  extends AppTest{
 
 //    test for COUNT(dataSources)
       String countDataSources =
-              App.executor.eval("g.V().has('Metadata.Type.Object.Data_Source', eq('Object.Data_Source'))\n" +
+              App.executor.eval("App.g.V().has('Metadata.Type.Object.Data_Source', eq('Object.Data_Source'))\n" +
                       ".count().next().toString()").get().toString();
       assertEquals("8", countDataSources);
 
@@ -227,6 +199,12 @@ public class AppTest2  extends AppTest{
                       ".count().next().toString()").get().toString();
       // expecting 1 less Event.Ingestion because "sharepoint" is the Data Source for the Data Sources
       assertEquals("7", countEventIngestions);
+
+      String countObjectDataPolicy =
+          App.executor.eval("App.g.V().has('Metadata.Type.Object.Data_Policy', eq('Object.Data_Policy'))\n" +
+              ".count().next().toString()").get().toString();
+      // expecting 1 less Event.Ingestion because "sharepoint" is the Data Source for the Data Sources
+      assertEquals("7", countObjectDataPolicy);
 
     } catch (ExecutionException e) {
       e.printStackTrace();
@@ -268,7 +246,7 @@ public class AppTest2  extends AppTest{
                       ".next().id().toString()").get().toString();
       String comidas2ConnectionsQuery = "App.g.V(\"" + userId3 +"\").bothE().count().next().toString()";
       String comidas2Connections = App.executor.eval(comidas2ConnectionsQuery).get().toString();
-      assertEquals(comidas2Connections,"4");
+      assertEquals("5",comidas2Connections);
 
 //    test COUNT(Edges) for Object.Email_Address
       String userId4 =
@@ -296,23 +274,87 @@ public class AppTest2  extends AppTest{
 
   }
 
-  public void csvTestUtil(String csvFile,  String ruleName) throws InterruptedException {
+
+  @Test
+  public void testSQL() {
+    Gson gson = new Gson();
+    RecordRequest req = gson.fromJson(jsonReq, RecordRequest.class);
+
+    req.search.vid = "#-1:-1";
+    req.search.relationship = "has_server";
+    req.search.direction = "->";
+
+    String sqlCount = req.getSQL(true);
+
+    String expectedSqlCount = "SELECT COUNT(*) FROM (SELECT EXPAND(OUT('has_server')) FROM #-1:-1)";
+
+    assertEquals(expectedSqlCount, sqlCount);
+
+    String sql = req.getSQL(false);
+    String expectedSql = "SELECT @rid as id,`Event.Data_Breach.Description`,`Event.Data_Breach.Impact`  FROM (SELECT EXPAND(OUT('has_server')) FROM #-1:-1  WHERE (( `Event.Data_Breach.Impact`  containsText  'OS' ) ) ) ORDER BY `Event.Data_Breach.Impact` DESC SKIP 0 LIMIT 100";
+    assertEquals(expectedSql, sql);
+
+    req.search.direction = "<-";
+    sqlCount = req.getSQL(true);
+
+    expectedSqlCount = "SELECT COUNT(*) FROM (SELECT EXPAND(IN('has_server')) FROM #-1:-1)";
+    assertEquals(expectedSqlCount, sqlCount);
+
+    req.search.vid = null;
+    req.search.direction = null;
+    req.search.relationship = null;
+    sqlCount = req.getSQL(true);
+    expectedSqlCount = "SELECT COUNT(*)  FROM `" +
+        req.dataType +
+        "`  WHERE (( `Event.Data_Breach.Impact`  containsText  'OS' ) ) ";
+    assertEquals(expectedSqlCount, sqlCount);
+
+
+  }
+
+  @Test
+  public void testSimpleGremlin() throws InterruptedException {
+
+
+    String res = null;
+    try {
+      res = App.executor.eval("App.g.V().hasNext();").get().toString();
+      System.out.println(res);
+
+    } catch (ExecutionException e) {
+      e.printStackTrace();
+      assertNull(e);
+
+    }
+
+  }
+
+  @Test
+  public void testPhase1CsvDemoIngestion() throws InterruptedException {
+
 
     String res = null;
     try {
       Path resourceDirectory = Paths.get(".");
       String pwdAbsolutePath = resourceDirectory.toFile().getAbsolutePath();
-      String csvFilePath = Paths.get(pwdAbsolutePath, "src", "test", "resources", csvFile).toString();
+      String csvPath = Paths.get(pwdAbsolutePath, "src", "test", "resources", "phase1.csv").toString();
 
-      String csvString = FileUtils.readFileToString(new File(csvFilePath), "UTF-8");
+      Reader in = new FileReader(csvPath);
+      Iterable<CSVRecord> records = CSVFormat.RFC4180.withFirstRecordAsHeader().parse(in);
 
-      IngestionCSVFileRequest  req = new IngestionCSVFileRequest();
-      req.csvBase64 = Base64.getEncoder().encodeToString(csvString.getBytes());
-      req.ruleName = ruleName;
+      List<Map> listOfMaps = new ArrayList<>();
+      for (CSVRecord record : records) {
+        Map recMap = record.toMap();
+        recMap.put("currDate", new Date().toString());
+        listOfMaps.add(recMap);
+      }
 
-      Ingestion ingestion = new Ingestion();
 
-      res = ingestion.csvFile(req);
+      Map<String, Object> bindings = new HashMap() {{
+        put("listOfMaps", listOfMaps);
+      }};
+
+      res = App.executor.eval(phase1QueryStr, bindings).get().toString();
       System.out.println(res);
 
     } catch (ExecutionException | IOException e) {
@@ -320,4 +362,166 @@ public class AppTest2  extends AppTest{
       assertNull(e);
     }
   }
+
+  @Test
+  public void testTotvsProtheusSa1Clientes() throws InterruptedException {
+
+    jsonTestUtil("totvs1.json", "$.objs", "totvs_protheus_sa1_clientes");
+//    jsonTestUtil("totvs2.json", "$.objs", "totvs_protheus_sa1_clientes");
+
+    try {
+      String userId =
+          App.executor.eval("App.g.V().has('Person.Natural.Full_Name',eq('JONAS LEO BATISTA')).next().id().toString()")
+              .get().toString();
+
+      Resource res = new Resource();
+      GremlinRequest gremlinReq = new GremlinRequest();
+      gremlinReq.setGremlin(
+          "App.g.V().has('Person.Natural.Full_Name',eq('JONAS LEO BATISTA')).next().id().toString()");
+      JsonObject obj = JsonParser.parseString(res.gremlinQuery(gson.toJson(gremlinReq))).getAsJsonObject();
+      String stringifiedOutput = gson.toJson(obj);
+      String res2;
+      try {
+        HttpClient client = HttpClients.createMinimal();
+
+        HttpPost request = new HttpPost(URI.create("http://localhost:3001/home/gremlin"));
+
+        request.setHeader("Content-Type", "application/json");
+        request.setHeader("Accept", "application/json");
+
+        StringEntity data = new StringEntity(gson.toJson(gremlinReq));
+//        SerializableEntity data = new SerializableEntity(req);
+
+        request.setEntity(data);
+
+
+        res2 = IOUtils.toString(client.execute(request).getEntity().getContent());
+        System.out.println(res2);
+
+
+        HttpGet getRequest = new HttpGet(URI.create("http://localhost:3001/home/hello"));
+
+//        request.setHeader("Content-Type", "application/json");
+//        request.setHeader("Accept", "application/json");
+
+//        StringEntity getData = new StringEntity(gson.toJson(req));
+//        SerializableEntity data = new SerializableEntity(req);
+
+//        request.setEntity(data);
+
+
+        res2 = IOUtils.toString(client.execute(getRequest).getEntity().getContent());
+
+
+        System.out.println(res2);
+
+      } catch (IOException e) {
+        e.printStackTrace();
+        assertNull(e);
+      }
+
+
+      assertEquals(userId, ((obj.get("result").getAsJsonObject())
+          .getAsJsonObject().get("data").getAsJsonObject()).get("@value").getAsJsonArray().get(0).getAsString());
+
+      Map<String, Object> bindings = new HashMap() {{
+        put("pg_id", userId);
+        put("pg_templateText", "eyUgc2V0IHBvc3NpYmxlTWF0Y2hlcyA9IHB2OnBvc3NpYmxlTWF0Y2hlcyhjb250ZXh0LmlkLCd7Ik9iamVjdC5FbWFpbF9BZGRyZXNzIjogMTAuNSwgIk9iamVjdC5JZGVudGl0eV9DYXJkIjogOTAuNSwgIkxvY2F0aW9uLkFkZHJlc3MiOiAxMC4xLCAiT2JqZWN0LlBob25lX051bWJlciI6IDEuMCwgIk9iamVjdC5TZW5zdGl2ZV9EYXRhIjogMTAuMCwgIk9iamVjdC5IZWFsdGgiOiAxLjAsICJPYmplY3QuQmlvbWV0cmljIjogNTAuMCAsICJPYmplY3QuSW5zdXJhbmNlX1BvbGljeSI6IDEuMH0nKSAlfQp7JSBzZXQgbnVtTWF0Y2hlcyA9IHBvc3NpYmxlTWF0Y2hlcy5zaXplKCkgJX0KeyUgaWYgbnVtTWF0Y2hlcyA9PSAwICV9Cnt7IGNvbnRleHQuUGVyc29uX05hdHVyYWxfRnVsbF9OYW1lfX0gw6kgbyDDum5pY28gcmVnaXN0cm8gbm8gc2lzdGVtYS4KCnslIGVsc2UgJX0Ke3sgY29udGV4dC5QZXJzb25fTmF0dXJhbF9GdWxsX05hbWV9fSBDb3JyZXNwb25kZSBwb3RlbmNpYWxtZW50ZSBhIHt7IG51bU1hdGNoZXMgfX0gcmVnaXN0cm97JS0gaWYgbnVtTWF0Y2hlcyAhPSAxIC0lfXN7JSBlbmRpZiAlfS4KCgoKICB7eyAiPHRhYmxlIHN0eWxlPSdtYXJnaW46IDVweCc+PHRyIHN0eWxlPSdib3JkZXI6IDFweCBzb2xpZCAjZGRkZGRkO3RleHQtYWxpZ246IGxlZnQ7cGFkZGluZzogOHB4Oyc+PHRoIHN0eWxlPSdib3JkZXI6IDFweCBzb2xpZCAjZGRkZGRkO3RleHQtYWxpZ246IGxlZnQ7cGFkZGluZzogOHB4Oyc+VGl0dWxhcjwvdGg+PHRoIHN0eWxlPSdib3JkZXI6IDFweCBzb2xpZCAjZGRkZGRkO3RleHQtYWxpZ246IGxlZnQ7cGFkZGluZzogOHB4Oyc+UGVyY2VudHVhbDwvdGg+PHRoIHN0eWxlPSdib3JkZXI6IDFweCBzb2xpZCAjZGRkZGRkO3RleHQtYWxpZ246IGxlZnQ7cGFkZGluZzogOHB4Oyc+UHJvcHJpZWRhZGVzIGVtIENvbXVtPC90aD48L3RyPiIgfX0KICB7JSBmb3IgaXRlbSBpbiBwb3NzaWJsZU1hdGNoZXMuZW50cnlTZXQoKSAlfQogIHt7ICAiPHRyIHN0eWxlPSdib3JkZXI6IDFweCBzb2xpZCAjZGRkZGRkO3RleHQtYWxpZ246IGxlZnQ7cGFkZGluZzogOHB4Oyc+PHRkIHN0eWxlPSdib3JkZXI6IDFweCBzb2xpZCAjZGRkZGRkO3RleHQtYWxpZ246IGxlZnQ7cGFkZGluZzogOHB4Oyc+JXM8L3RkPjx0ZCBzdHlsZT0nYm9yZGVyOiAxcHggc29saWQgI2RkZGRkZDt0ZXh0LWFsaWduOiBsZWZ0O3BhZGRpbmc6IDhweDsnPiUuMmYlJTwvdGQ+PHRkIHN0eWxlPSdib3JkZXI6IDFweCBzb2xpZCAjZGRkZGRkO3RleHQtYWxpZ246IGxlZnQ7cGFkZGluZzogOHB4Oyc+JXM8L3RkPiIgfCBmb3JtYXQgKGl0ZW0ua2V5LlBlcnNvbl9OYXR1cmFsX0Z1bGxfTmFtZSAsIGl0ZW0udmFsdWUgKiAxMDAuMCwgaXRlbS5rZXkuTGFiZWxzX0Zvcl9NYXRjaCApIH19CiAgeyUgZW5kZm9yICV9CiAge3sgIjwvdGFibGU+IiB9fQoKeyUgZW5kaWYgJX0=");
+      }};
+
+      String report = App.executor.eval("renderReportInBase64(pg_id,pg_templateText)", bindings).get().toString();
+      System.out.println(report);
+
+    } catch (ExecutionException e) {
+      e.printStackTrace();
+      assertNull(e);
+
+    }
+
+
+  }
+
+  @Test
+  public void testPloomesClientes() throws InterruptedException {
+
+    jsonTestUtil("ploomes1.json", "$.value", "ploomes_clientes");
+//    jsonTestUtil("ploomes1.json", "$.value", "ploomes_clientes");
+
+    try {
+      String userId =
+          App.executor.eval("App.g.V().has('Person.Organisation.Short_Name',eq('Pessoa Nova5')).next().id().toString()")
+              .get().toString();
+
+      Resource res = new Resource();
+      GremlinRequest gremlinReq = new GremlinRequest();
+      gremlinReq.setGremlin(
+          "App.g.V().has('Person.Organisation.Short_Name',eq('Pessoa Nova5')).next().id().toString()");
+      JsonObject obj = JsonParser.parseString(res.gremlinQuery(gson.toJson(gremlinReq))).getAsJsonObject();
+//      Gson gson = new Gson();
+      String stringifiedOutput = gson.toJson(obj);
+      String res2;
+      try {
+        HttpClient client = HttpClients.createMinimal();
+
+        HttpPost request = new HttpPost(URI.create("http://localhost:3001/home/gremlin"));
+
+        request.setHeader("Content-Type", "application/json");
+        request.setHeader("Accept", "application/json");
+
+        StringEntity data = new StringEntity(gson.toJson(gremlinReq));
+//        SerializableEntity data = new SerializableEntity(req);
+
+        request.setEntity(data);
+
+
+        res2 = IOUtils.toString(client.execute(request).getEntity().getContent());
+        System.out.println(res2);
+
+
+        HttpGet getRequest = new HttpGet(URI.create("http://localhost:3001/home/hello"));
+
+//        request.setHeader("Content-Type", "application/json");
+//        request.setHeader("Accept", "application/json");
+
+//        StringEntity getData = new StringEntity(gson.toJson(req));
+//        SerializableEntity data = new SerializableEntity(req);
+
+//        request.setEntity(data);
+
+
+        res2 = IOUtils.toString(client.execute(getRequest).getEntity().getContent());
+
+
+        System.out.println(res2);
+
+      } catch (IOException e) {
+        e.printStackTrace();
+        assertNull(e);
+      }
+
+
+//      String b = "DeBugger Stop!";
+
+
+      assertEquals(userId, (((obj.get("result").getAsJsonObject())
+          .getAsJsonObject().get("data")).getAsJsonObject().get("@value")).getAsJsonArray().get(0).getAsString());
+
+      Map<String, Object> bindings = new HashMap() {{
+        put("pg_id", userId);
+        put("pg_templateText", "eyUgc2V0IHBvc3NpYmxlTWF0Y2hlcyA9IHB2OnBvc3NpYmxlTWF0Y2hlcyhjb250ZXh0LmlkLCd7Ik9iamVjdC5FbWFpbF9BZGRyZXNzIjogMTAuNSwgIk9iamVjdC5JZGVudGl0eV9DYXJkIjogOTAuNSwgIkxvY2F0aW9uLkFkZHJlc3MiOiAxMC4xLCAiT2JqZWN0LlBob25lX051bWJlciI6IDEuMCwgIk9iamVjdC5TZW5zdGl2ZV9EYXRhIjogMTAuMCwgIk9iamVjdC5IZWFsdGgiOiAxLjAsICJPYmplY3QuQmlvbWV0cmljIjogNTAuMCAsICJPYmplY3QuSW5zdXJhbmNlX1BvbGljeSI6IDEuMH0nKSAlfQp7JSBzZXQgbnVtTWF0Y2hlcyA9IHBvc3NpYmxlTWF0Y2hlcy5zaXplKCkgJX0KeyUgaWYgbnVtTWF0Y2hlcyA9PSAwICV9Cnt7IGNvbnRleHQuUGVyc29uX05hdHVyYWxfRnVsbF9OYW1lfX0gw6kgbyDDum5pY28gcmVnaXN0cm8gbm8gc2lzdGVtYS4KCnslIGVsc2UgJX0Ke3sgY29udGV4dC5QZXJzb25fTmF0dXJhbF9GdWxsX05hbWV9fSBDb3JyZXNwb25kZSBwb3RlbmNpYWxtZW50ZSBhIHt7IG51bU1hdGNoZXMgfX0gcmVnaXN0cm97JS0gaWYgbnVtTWF0Y2hlcyAhPSAxIC0lfXN7JSBlbmRpZiAlfS4KCgoKICB7eyAiPHRhYmxlIHN0eWxlPSdtYXJnaW46IDVweCc+PHRyIHN0eWxlPSdib3JkZXI6IDFweCBzb2xpZCAjZGRkZGRkO3RleHQtYWxpZ246IGxlZnQ7cGFkZGluZzogOHB4Oyc+PHRoIHN0eWxlPSdib3JkZXI6IDFweCBzb2xpZCAjZGRkZGRkO3RleHQtYWxpZ246IGxlZnQ7cGFkZGluZzogOHB4Oyc+VGl0dWxhcjwvdGg+PHRoIHN0eWxlPSdib3JkZXI6IDFweCBzb2xpZCAjZGRkZGRkO3RleHQtYWxpZ246IGxlZnQ7cGFkZGluZzogOHB4Oyc+UGVyY2VudHVhbDwvdGg+PHRoIHN0eWxlPSdib3JkZXI6IDFweCBzb2xpZCAjZGRkZGRkO3RleHQtYWxpZ246IGxlZnQ7cGFkZGluZzogOHB4Oyc+UHJvcHJpZWRhZGVzIGVtIENvbXVtPC90aD48L3RyPiIgfX0KICB7JSBmb3IgaXRlbSBpbiBwb3NzaWJsZU1hdGNoZXMuZW50cnlTZXQoKSAlfQogIHt7ICAiPHRyIHN0eWxlPSdib3JkZXI6IDFweCBzb2xpZCAjZGRkZGRkO3RleHQtYWxpZ246IGxlZnQ7cGFkZGluZzogOHB4Oyc+PHRkIHN0eWxlPSdib3JkZXI6IDFweCBzb2xpZCAjZGRkZGRkO3RleHQtYWxpZ246IGxlZnQ7cGFkZGluZzogOHB4Oyc+JXM8L3RkPjx0ZCBzdHlsZT0nYm9yZGVyOiAxcHggc29saWQgI2RkZGRkZDt0ZXh0LWFsaWduOiBsZWZ0O3BhZGRpbmc6IDhweDsnPiUuMmYlJTwvdGQ+PHRkIHN0eWxlPSdib3JkZXI6IDFweCBzb2xpZCAjZGRkZGRkO3RleHQtYWxpZ246IGxlZnQ7cGFkZGluZzogOHB4Oyc+JXM8L3RkPiIgfCBmb3JtYXQgKGl0ZW0ua2V5LlBlcnNvbl9OYXR1cmFsX0Z1bGxfTmFtZSAsIGl0ZW0udmFsdWUgKiAxMDAuMCwgaXRlbS5rZXkuTGFiZWxzX0Zvcl9NYXRjaCApIH19CiAgeyUgZW5kZm9yICV9CiAge3sgIjwvdGFibGU+IiB9fQoKeyUgZW5kaWYgJX0=");
+      }};
+
+      String report = App.executor.eval("renderReportInBase64(pg_id,pg_templateText)", bindings).get().toString();
+      System.out.println(report);
+
+    } catch (ExecutionException e) {
+      e.printStackTrace();
+      assertNull(e);
+
+    }
+
+
+  }
+
+
 }
