+++ src/test/resources/ploomes1.json
@@ -253,7 +253,7 @@
     },
     {
       "Id": 13181014,
-      "TypeId": 2,
+      "TypeId": 1,
       "Name": "Pessoa Nova5",
       "LegalName": null,
       "Register": null,
