+++ src/test/resources/totvs1.json
@@ -136,6 +136,73 @@
       "A1_ENDREC ": "",
       "id": "13181011",
       "A1_LC     ": "0"
+    },
+    {
+      "A1_PAGATR ": "0",
+      "A1_BAIRROC": "BAIRRO 2",
+      "A1_COMPENT": "",
+      "firstUpdate": "2021-01-01T07:59:59",
+      "lastUpdate": "2021-05-19T08:00:00",
+      "A1_TITPROT": "0",
+      "A1_BAIRROE": "BAIRRO 2",
+      "A1_CGC    ": "06963123456",
+      "A1_CEPC   ": "13290000",
+      "A1_CEP    ": "87698453",
+      "A1_CODMUNE": "",
+      "A1_ATR    ": "0",
+      "A1_IBGE   ": "",
+      "A1_CEPE   ": "13290000",
+      "A1_BAIRRO ": "BAIRRO 2",
+      "A1_ESTADO ": "SP",
+      "A1_DDD    ": "19",
+      "A1_INSCR  ": "421075333333",
+      "A1_CODMUN ": "",
+      "A1_PFISICA": "",
+      "A1_DDI    ": "55",
+      "A1_EMAIL  ": "JONAS@COMIDA2.COM.BR",
+      "A1_CLASSE ": "",
+      "A1_SALFIN ": "0",
+      "A1_END    ": "RUA NUMERO 2, 1",
+      "A1_ENDENT ": "RUA NUMERO 2, 1",
+      "A1_REGIAO ": "005",
+      "A1_EST    ": "SP",
+      "A1_ESTC   ": "SP",
+      "A1_PAIS   ": "",
+      "A1_LCFIN  ": "0",
+      "A1_METR   ": "0",
+      "A1_RG     ": "100984563",
+      "A1_MUN    ": "BOTUCATU",
+      "A1_ESTE   ": "SP",
+      "A1_SALFINM": "0",
+      "A1_MUNC   ": "BOTUCATU",
+      "A1_CODTER ": "",
+      "A1_ENDCOB ": "RUA NUMERO 2, 1",
+      "A1_MUNE   ": "BOTUCATU",
+      "A1_FAX    ": "12345678",
+      "A1_NIF    ": "",
+      "A1_NREDUZ ": "CIDADE GRANDE",
+      "A1_RISCO  ": "E",
+      "A1_HPAGE  ": "",
+      "A1_CODLOC ": "",
+      "A1_INSCRUR": "",
+      "A1_CEINSS ": "",
+      "A1_CHQDEVO": "0",
+      "A1_DSCREG ": "",
+      "A1_TEL    ": "22222222",
+      "A1_COMPLEM": "",
+      "A1_COD_MUN": "27306",
+      "A1_CXPOSTA": "",
+      "A1_NOME   ": "JONAS LEO BATISTA",
+      "A1_CONTATO": "JONAS",
+      "A1_INSCRM ": "",
+      "A1_TELEX  ": "",
+      "A1_CODPAIS": "01058",
+      "A1_DTULTIT": "\/  \/",
+      "A1_DTNASC ": "\/  \/",
+      "A1_DTULCHQ": "\/  \/",
+      "A1_ENDREC ": "",
+      "id": "13181099",
+      "A1_LC     ": "0"
     }
   ]
 }
